requirements.txt
requests==2.26.0
Flask==2.0.2
pycryptodome
